up: fix some error on run v3 examples
