+++ _examples/cliapp/main.go
@@ -46,7 +46,9 @@ func main() {
 	app.Add(cmd.ExampleCommand())
 	app.Add(cmd.DaemonRunCommand())
 	app.Add(cmd.EnvInfoCommand())
-	app.Add(cmd.GitCommand(), cmd.GitPullMulti)
+
+	app.Add(cmd.GitCmd)
+
 	app.Add(cmd.ColorCommand(), cmd.EmojiDemoCmd())
 	app.Add(cmd.ShowDemoCommand(), cmd.ProgressDemoCmd(), cmd.SpinnerDemoCmd(), cmd.InteractDemoCommand())
 	app.Add(builtin.GenEmojiMapCommand())
