+++ _examples/cmd/cmd.go
@@ -1,0 +1,5 @@
+package cmd
+
+func init() {
+	GitCmd.Add(GitInfoCommand(), GitPullMulti)
+}
