+++ _examples/cmd/git.go
@@ -1,0 +1,8 @@
+package cmd
+
+import "github.com/gookit/gcli/v3"
+
+var GitCmd = &gcli.Command{
+	Name: "git",
+	Desc: "git usage",
+}
