+++ _examples/cmd/git_pull_multi.go
@@ -16,9 +16,9 @@ import (
 
 // GitPullMulti use git pull for update multi project
 var GitPullMulti = &gcli.Command{
-	Name:    "git:pull",
+	Name:    "pull",
 	Desc:    "use git pull for update multi project",
-	Aliases: []string{"git-pull"},
+	Aliases: []string{"pul"},
 	Config: func(c *gcli.Command) {
 		c.AddArg(
 			"basePath",
