+++ internal/help_tpl.go
@@ -1,2 +1,1 @@
 package internal
-
